Fix P0 integer-overflow path in image-buffer allocation (+P2 short-write cleanup)
- render.c/image.c: cast every operand to size_t so w*h*3 is computed in
  size_t, never in 32-bit int (overflowed for dimensions > ~46341,
  producing an undersized buffer and a heap overflow during render).
- main.c parse_int_arg: reject strtol ERANGE and values outside
  [min, INT_MAX] so --width 2147483648 can no longer wrap; per-flag min
  (1 for dimensions/samples/depth, 0 for --threads = auto).
- main.c: add a MAX_PIXELS budget cap so absurd-but-valid dimensions
  (e.g. 50000x50000 = 7.5 GB) are rejected up front instead of attempting
  a multi-GB allocation and a runaway render.
- image.c (P2): remove(path) on a short PPM write so a truncated, corrupt
  image is never left on disk mistakable for a valid one.

+++ src/render.c
@@ -114,7 +114,9 @@ unsigned char *render_scene(const Scene *scene, const RenderConfig *cfg) {
     int h = cfg->height;
     int n = cfg->samples < 1 ? 1 : cfg->samples;
 
-    unsigned char *pixels = malloc((size_t)w * h * 3);
+    /* Cast every operand to size_t so the product is computed in size_t
+     * arithmetic, never in 32-bit int (which would overflow for large w*h). */
+    unsigned char *pixels = malloc((size_t)w * (size_t)h * 3);
     if (!pixels) {
         return NULL;
     }
